# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A simple, fast URL shortener API built with Go + Gin + PostgreSQL.",
        "title": "Linkas API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Galan Jabal",
            "url": "https://github.com/galanjabal3",
            "email": "galanjabal3@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/health": {
            "get": {
                "description": "Check if the API is running",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/qrcode/{slug}": {
            "get": {
                "description": "Generate QR code image for a short URL",
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Get QR Code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short URL slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/shorten": {
            "post": {
                "description": "Create a new short URL from a long URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Shorten URL",
                "parameters": [
                    {
                        "description": "URL to shorten",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ShortenRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.ShortenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/stats/{slug}": {
            "get": {
                "description": "Get click statistics for a short URL",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Get URL statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short URL slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.StatsResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/{slug}": {
            "delete": {
                "description": "Delete a short URL by slug",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Delete short URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short URL slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/{slug}": {
            "get": {
                "description": "Redirect short URL to original destination",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "urls"
                ],
                "summary": "Redirect to original URL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Short URL slug",
                        "name": "slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "301": {
                        "description": "Moved Permanently"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Click": {
            "type": "object",
            "properties": {
                "clicked_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "ip_address": {
                    "type": "string"
                },
                "referrer": {
                    "type": "string"
                },
                "url_id": {
                    "type": "integer"
                },
                "user_agent": {
                    "type": "string"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "model.ShortenRequest": {
            "type": "object",
            "required": [
                "expires_in",
                "url"
            ],
            "properties": {
                "custom_slug": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "model.ShortenResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "original_url": {
                    "type": "string"
                },
                "short_url": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
        "model.StatsResponse": {
            "type": "object",
            "properties": {
                "click_count": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "original_url": {
                    "type": "string"
                },
                "recent_clicks": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Click"
                    }
                },
                "slug": {
                    "type": "string"
                }
            }
        }
    }
}